change dÃata type Integer in Double

+++ src/main/java/com/technology/apigateway/database/entity/OperatorEstPayment.java
@@ -52,13 +52,13 @@ public class OperatorEstPayment {
     private Date duedate;
 
     @Column(name = "PRINNML")
-    private Integer prinnml;
+    private Double prinnml;
 
     @Column(name = "ESTINT")
-    private Integer estint;
+    private Double estint;
 
     @Column(name = "ESTPRIN")
-    private Integer estprin;
+    private Double estprin;
 
     @Column(name = "DUENO")
     private Integer dueno;
